added id validation per solution

+++ routes/notes.js
@@ -1,11 +1,10 @@
 'use strict';
 
 const express = require('express');
-
 const mongoose = require('mongoose');
-const { MONGODB_URI } = require('../config');
 
 const Note = require('../models/note');
+
 
 const router = express.Router();
 
@@ -13,9 +12,11 @@ const router = express.Router();
 router.get('/', (req, res, next) => {
   const searchTerm = req.query.searchTerm;
   let filter = {};
+  
   if (searchTerm) {
     filter.title = { $regex: searchTerm, $options: 'i'};
   }
+  
   Note.find(filter).sort({updatedAt: 'desc'})
     .then(results => {
       res.json(results);
@@ -29,6 +30,7 @@ router.get('/', (req, res, next) => {
 /* ========== GET/READ A SINGLE ITEM ========== */
 router.get('/:id', (req, res, next) => {
   const id = req.params.id;
+  
   Note.findById(id)
     .then(results => {
       res.json(results);
@@ -37,22 +39,23 @@ router.get('/:id', (req, res, next) => {
       console.error(`ERROR: ${err.message}`);
       next(err);
     });
-  // console.log('Get a Note');
-  // res.json({ id: 1, title: 'Temp 1' });
 });
 
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/', (req, res, next) => {
   const { title, content } = req.body;
   const newNote = {
-    title: title,
-    content: content
+    title, content
+    // title: title,
+    // content: content
   };
+  
   if (!newNote.title) {
     const err = new Error('Missing `title` in request body');
     err.status = 400;
     return next(err);
   }
+  
   Note.create({
     title: newNote.title,
     content: newNote.content
@@ -74,22 +77,45 @@ router.put('/:id', (req, res, next) => {
     title: title,
     content: content
   };
-  Note.findByIdAndUpdate(id, { $set: {title: updateNote.title, content: updateNote.content} }, {new: true})
+  
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error('The `id` is not valid');
+    err.status = 400;
+    return next(err);
+  }
+  
+  if (!updateNote.title) {
+    const err = new Error('Missing `title` in request body');
+    err.status = 400;
+    return next(err);
+  }
+  
+  // Note.findByIdAndUpdate(id, { $set: {title: updateNote.title, content: updateNote.content} }, {new: true})
+  Note.findByIdAndUpdate(id, updateNote, {new: true})
     .then(results => {
-      res.json(results);
+      if (results){
+        res.json(results);
+      } else {
+        next();
+      }
     })
     .catch(err => {
       console.error(`ERROR: ${err.message}`);
       next(err);
     });
 });
-// console.log('Update a Note');
-// res.json({ id: 1, title: 'Updated Temp 1' });
 
 
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/:id', (req, res, next) => {
   const deleteId = req.params.id;
+  
+  if (!mongoose.Types.ObjectId.isValid(deleteId)) {
+    const err = new Error('The `id` is not valid');
+    err.status = 400;
+    return next(err);
+  }
+  
   Note.findByIdAndRemove(deleteId)
     .then(() => {
       res.sendStatus(204);
@@ -99,7 +125,5 @@ router.delete('/:id', (req, res, next) => {
       next(err);
     });
 });
-// console.log('Delete a Note');
-// res.status(204).end();
 
 module.exports = router;
